Feature/matchgame view

+++ app/Http/Controllers/Auth/TicketViewController.php
@@ -7,57 +7,87 @@ use App\Models\Matchgame;
 use App\Models\TeamPlayingMatch;
 use App\Models\Ticket;
 use App\Models\Zone;
-use Illuminate\Foundation\Auth\AuthenticatesUsers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Validation\ValidationException;
+use App\Http\Controllers\Validation\DataValidator;
 
 class TicketViewController extends HomeController
 {
     public function index()
     {
-        $tickets = Ticket::all();
-        $zones = Zone::all();
-        $matchgames = MatchGame::all();
-        $team_playing_match = TeamPlayingMatch::all();
-        return view('tickets',
-                ['tickets' => $tickets,
-                'zones'=>$zones,
-                'matchgames'=>$matchgames,
-                'team_playing_match'=>$team_playing_match
-                ]
+        $tickets = Ticket::orderBy('id')->paginate(20);
+        $matchgames = MatchGame::withTrashed()->get();
+        return view('ticketsIndex', compact('tickets'),
+                compact('matchgames')
         );
     }
 
-    public function delete(Ticket $ticket)
+    public function delete($ticketId)
     {
+        request()->merge(['ticketId' => request()->route('ticketId')]);
+
+        $validator = DataValidator::validateTicketID(request()->all());
+
+        if($validator->fails()){
+            return redirect()->back()->withErrors([$validator->errors()->first()])->withInput();
+        }
+
+        $ticket = Ticket::find($ticketId);
+
+        if(!$ticket)
+            return redirect()->back()->withErrors(["Ticket not found"])->withInput();
+        
         $ticket->delete();
+        
         return redirect()->back()->with('success', 'Ticket deleted successfully.');
     }
 
-    public function update(Request $request, Ticket $ticket)
+    public function update(Request $request, $ticketId)
     {
-        $ticket->zone_id = $request->zone_id;
-        $ticket->base_price = $request->base_price;
-        $ticket->matchgame_id = $request->matchgame_id;
+        request()->merge(['ticketId' => request()->route('ticketId')]);
+
+        $validator = DataValidator::validate(request()->all(), [
+            'ticketId' => 'required|exists:tickets,id',
+            'price' => 'required|integer|min:1|max:10000',
+            'zoneId' => 'required|exists:zones,id',
+            'category' => 'string',
+        ]);
+
+        if($validator->fails()){
+            return redirect()->back()->withErrors([$validator->errors()->first()])->withInput();
+        }
+
+        $ticket = Ticket::find($ticketId);
+
+        $ticket->base_price = $request->price;
+        $ticket->zone_id = $request->zoneId;
+        $ticket->category = $request->category;
+
         $ticket->save();
-        return redirect()->back()->with('success', 'Ticket updated successfully.');
+
+        return redirect("/tickets/index")->with('success', 'Matchgame updated successfully.');
     }
 
     public function store(Request $request)
     {
+        $validator = DataValidator::validate(request()->all(), [
+            'matchgameId' => 'required|exists:matchgames,id',
+            'zoneId' => 'required|exists:zones,id',
+            'price' => 'required|integer|min:1|max:10000',
+            'category' => 'string',
+        ]);
 
-        $ticket = new Ticket;
-        $ticket->zone_id = $request->zone_id;
-        $ticket->base_price = $request->base_price;
-        $ticket->matchgame_id = $request->matchgame_id;
-
-        try {
-            $ticket->save();
-            return redirect()->back()->with('success', 'Ticket created successfully!');
-        } catch (\Exception $e) {
-            return redirect()->back()->with('error', 'Ticket creation failed: ' . $e->getMessage());
+        if($validator->fails()){
+            return redirect()->back()->withErrors([$validator->errors()->first()])->withInput();
         }
+        
+        $ticket = Ticket::factory()->state(['category' => $request->category , 'base_price' => $request->price])->create();
+        $ticket->zone_id = $request->zoneId;
+
+        $ticket->save();
+
+        return redirect("/tickets/index")->with('success', 'Ticket created successfully.');
     }
 
     public static function getMatchGameData() {
@@ -113,5 +143,50 @@ class TicketViewController extends HomeController
     public static function getZone($zones,$zoneId){
         $zone = $zones->where('id',$zoneId)->first();
         return $zone;
+    }
+
+    public function showCreatePage(){
+
+        $matchgames = MatchGame::all();
+
+        return view('ticketsCreate',compact('matchgames'));
+    }
+
+    public function showEditPage($ticketId){
+        request()->merge(['ticketId' => request()->route('ticketId')]);
+
+        $validator = DataValidator::validateTicketID(request()->all());
+
+        if($validator->fails()){
+            return redirect()->back()->withErrors([$validator->errors()->first()])->withInput();
+        }
+
+        $ticket = Ticket::find($ticketId);
+
+        if(!$ticket)
+            return redirect()->back()->withErrors(["Ticket not found"])->withInput();
+
+        $stadiumZones = $ticket->matchgame->stadium->zones;
+        return view('ticketsEdit',
+                [
+                    'ticket' => $ticket,
+                    'stadiumZones' => $stadiumZones,
+                ]
+        );
+    }
+
+    public function matchgameTickets(Request $request){
+
+        $validator = DataValidator::validateMatchgameID($request->all());
+
+        if($validator->fails()){
+            return redirect()->back()->withErrors([$validator->errors()->first()])->withInput();
+        }
+        
+        $tickets = Ticket::orderBy('id')->where('matchgame_id', $request->matchgameId)->withTrashed()->paginate(20);
+        
+        $matchgames = Matchgame::withTrashed()->get();
+        
+        return view('ticketsIndex', compact('tickets'), compact('matchgames'));
     }
 }
